Split tool workspace surfaces
