+++ tests/upscale-workspace-architecture.test.ts
@@ -13,6 +13,7 @@ const pricingHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks
 const previewScrollerHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscalePreviewScroller.ts');
 const previewStateHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscalePreviewState.ts');
 const recentJobsHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscaleRecentJobs.ts');
+const recentActionsHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscaleRecentActions.ts');
 const sourceMediaHookPath = join(root, 'frontend/src/components/tools/upscale/_hooks/useUpscaleSourceMedia.ts');
 const controlsPath = join(root, 'frontend/src/components/tools/upscale/_components/upscale-workspace-controls.tsx');
 const inputPanelsPath = join(root, 'frontend/src/components/tools/upscale/_components/UpscaleInputPanels.tsx');
@@ -32,6 +33,7 @@ test('upscale workspace delegates copy, local types, and helper logic', () => {
   assert.ok(existsSync(previewScrollerHookPath), 'upscale preview scroller should live in a colocated hook');
   assert.ok(existsSync(previewStateHookPath), 'upscale preview state should live in a colocated hook');
   assert.ok(existsSync(recentJobsHookPath), 'upscale recent jobs orchestration should live in a colocated hook');
+  assert.ok(existsSync(recentActionsHookPath), 'upscale recent actions should live in a colocated hook');
   assert.ok(existsSync(sourceMediaHookPath), 'upscale source media orchestration should live in a colocated hook');
   assert.ok(existsSync(controlsPath), 'upscale workspace controls should live in colocated components');
   assert.ok(existsSync(inputPanelsPath), 'upscale input panels should live in colocated components');
@@ -42,12 +44,12 @@ test('upscale workspace delegates copy, local types, and helper logic', () => {
 
   assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-copy'/, 'workspace should import upscale copy');
   assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-types'/, 'workspace should import upscale local types');
-  assert.match(workspaceSource, /from '\.\/upscale\/_lib\/upscale-workspace-helpers'/, 'workspace should import upscale helpers');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscaleLibraryAssets'/, 'workspace should import library hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscalePricingPreview'/, 'workspace should import pricing hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscalePreviewScroller'/, 'workspace should import preview scroller hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscalePreviewState'/, 'workspace should import preview state hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscaleRecentJobs'/, 'workspace should import recent jobs hook');
+  assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscaleRecentActions'/, 'workspace should import recent actions hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_hooks\/useUpscaleSourceMedia'/, 'workspace should import source media hook');
   assert.match(workspaceSource, /from '\.\/upscale\/_components\/UpscaleInputPanels'/, 'workspace should import input panels');
   assert.match(workspaceSource, /from '\.\/upscale\/_components\/UpscaleHeroSummaryCard'/, 'workspace should import hero summary card');
@@ -71,6 +73,12 @@ test('upscale workspace does not regain extracted ownership', () => {
   assert.doesNotMatch(workspaceSource, /useInfiniteJobs\(12, \{ surface: 'upscale' \}\)/, 'recent upscale feed belongs in useUpscaleRecentJobs');
   assert.doesNotMatch(workspaceSource, /getJobStatus/, 'recent job polling belongs in useUpscaleRecentJobs');
   assert.doesNotMatch(workspaceSource, /defaultGeneratedImageAppliedRef/, 'default generated image hydration belongs in useUpscaleRecentJobs');
+  assert.doesNotMatch(workspaceSource, /function buildRecentUpscaleResult/, 'recent result projection belongs in useUpscaleRecentActions');
+  assert.doesNotMatch(workspaceSource, /function resolveRecentSelectionWithSource/, 'recent detail hydration belongs in useUpscaleRecentActions');
+  assert.doesNotMatch(workspaceSource, /function selectRecentUpscale/, 'recent selection belongs in useUpscaleRecentActions');
+  assert.doesNotMatch(workspaceSource, /function handleRecentGroupAction/, 'recent grouped actions belong in useUpscaleRecentActions');
+  assert.doesNotMatch(workspaceSource, /authFetch/, 'recent job detail fetching belongs in useUpscaleRecentActions');
+  assert.doesNotMatch(workspaceSource, /formatCurrency/, 'recent price display belongs in useUpscaleRecentActions');
   assert.doesNotMatch(workspaceSource, /mediaTypeFromMime/, 'source media MIME detection belongs in useUpscaleSourceMedia');
   assert.doesNotMatch(workspaceSource, /new window\.Image/, 'source image dimension hydration belongs in useUpscaleSourceMedia');
   assert.doesNotMatch(workspaceSource, /function selectLibraryAsset\(/, 'library source selection belongs in useUpscaleSourceMedia');
@@ -98,7 +106,7 @@ test('upscale workspace does not regain extracted ownership', () => {
   assert.doesNotMatch(workspaceSource, /copy\.libraryUse/, 'library asset actions belong in UpscaleLibraryModal');
 
   const lineCount = workspaceSource.split('\n').length;
-  assert.ok(lineCount <= 620, `UpscaleWorkspace should stay below 620 lines after library modal extraction, got ${lineCount}`);
+  assert.ok(lineCount <= 480, `UpscaleWorkspace should stay below 480 lines after recent action extraction, got ${lineCount}`);
 });
 
 test('upscale helper modules expose the expected workspace contract', () => {
@@ -110,6 +118,7 @@ test('upscale helper modules expose the expected workspace contract', () => {
   const previewScrollerHookSource = readFileSync(previewScrollerHookPath, 'utf8');
   const previewStateHookSource = readFileSync(previewStateHookPath, 'utf8');
   const recentJobsHookSource = readFileSync(recentJobsHookPath, 'utf8');
+  const recentActionsHookSource = readFileSync(recentActionsHookPath, 'utf8');
   const sourceMediaHookSource = readFileSync(sourceMediaHookPath, 'utf8');
   const controlsSource = readFileSync(controlsPath, 'utf8');
   const inputPanelsSource = readFileSync(inputPanelsPath, 'utf8');
@@ -164,6 +173,11 @@ test('upscale helper modules expose the expected workspace contract', () => {
   assert.match(recentJobsHookSource, /useInfiniteJobs\(12, \{ surface: 'upscale' \}\)/, 'recent jobs hook should own the upscale feed');
   assert.match(recentJobsHookSource, /getJobStatus/, 'recent jobs hook should poll pending jobs');
   assert.match(recentJobsHookSource, /resolveGeneratedImageSource/, 'recent jobs hook should hydrate the default generated image source');
+  assert.match(recentActionsHookSource, /export function useUpscaleRecentActions/, 'recent actions hook should be exported');
+  assert.match(recentActionsHookSource, /buildRecentUpscaleResult/, 'recent actions hook should project selected jobs into results');
+  assert.match(recentActionsHookSource, /authFetch/, 'recent actions hook should hydrate detailed job state when needed');
+  assert.match(recentActionsHookSource, /saveRecentUpscale/, 'recent actions hook should own saving recent outputs');
+  assert.match(recentActionsHookSource, /handleRecentGroupAction/, 'recent actions hook should own grouped card actions');
   assert.match(sourceMediaHookSource, /export function useUpscaleSourceMedia/, 'source media hook should be exported');
   assert.match(sourceMediaHookSource, /uploadSourceFile/, 'source media hook should own source uploads');
   assert.match(sourceMediaHookSource, /mediaTypeFromMime/, 'source media hook should own MIME detection');
